+++ azure/flatcar-linux/kubernetes/bootstrap.tf
@@ -1,17 +1,12 @@
 # Kubernetes assets (kubeconfig, manifests)
 module "bootstrap" {
-  source = "git::https://github.com/poseidon/terraform-render-bootstrap.git?ref=c775b4de9a16ad1a94fef811f891c49169e7729f"
+  source = "git::https://github.com/poseidon/terraform-render-bootstrap.git?ref=33f8d2083cd2da5a18f954dd4f765b482d9b8046"
 
   cluster_name = var.cluster_name
   api_servers  = [format("%s.%s", var.cluster_name, var.dns_zone)]
   etcd_servers = formatlist("%s.%s", azurerm_dns_a_record.etcds.*.name, var.dns_zone)
 
-  networking = var.networking
-  # only effective with Calico networking
-  # we should be able to use 1450 MTU, but in practice, 1410 was needed
-  network_encapsulation = "vxlan"
-  network_mtu           = "1410"
-
+  networking            = var.networking
   pod_cidr              = var.pod_cidr
   service_cidr          = var.service_cidr
   daemonset_tolerations = var.daemonset_tolerations