+++ digital-ocean/flatcar-linux/kubernetes/variables.tf
@@ -67,7 +67,7 @@ variable "ssh_fingerprints" {
 
 variable "networking" {
   type        = string
-  description = "Choice of networking provider (flannel, calico, or cilium)"
+  description = "Choice of networking provider (flannel or cilium)"
   default     = "cilium"
 }