+++ google-cloud/fedora-coreos/kubernetes/variables.tf
@@ -116,7 +116,7 @@ variable "ssh_authorized_key" {
 
 variable "networking" {
   type        = string
-  description = "Choice of networking provider (flannel, calico, or cilium)"
+  description = "Choice of networking provider (flannel or cilium)"
   default     = "cilium"
 }